+++ apm-sniffer/apm-sdk-plugin/baidu-brpc-plugin/pom.xml
@@ -26,18 +26,20 @@
         <version>8.1.0-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
+    <packaging>jar</packaging>
 
     <artifactId>baidu-brpc-plugin</artifactId>
+    <name>baidu-brpc-plugin</name>
 
     <properties>
-        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+        <brpc-java.version>2.5.3</brpc-java.version>
     </properties>
 
     <dependencies>
         <dependency>
             <groupId>com.baidu</groupId>
             <artifactId>brpc-java</artifactId>
-            <version>2.5.3</version>
+            <version>${brpc-java.version}</version>
             <scope>provided</scope>
         </dependency>
     </dependencies>
